Все сделано, кроме вывода

+++ src/main/java/model/Discountable.java
@@ -1,5 +1,5 @@
 package model;
 
 public interface Discountable {
-    int getDiscount();
+    double getDiscount();
 }
